Extract fetch_and_structure_rewards function for dryer code
