+++ apps/explorer/lib/explorer/celo/util.ex
@@ -2,10 +2,19 @@ defmodule Explorer.Celo.Util do
   @moduledoc """
   Utilities for reading from Celo smart contracts
   """
+  import Ecto.Query,
+    only: [
+      from: 2
+    ]
 
   require Logger
-  alias Explorer.Celo.{AbiHandler, AddressCache}
+  alias Explorer.Celo.{AbiHandler, AddressCache, ContractEvents}
+  alias Explorer.Chain.{Block, CeloContractEvent}
+  alias Explorer.Repo
   alias Explorer.SmartContract.Reader
+
+  alias ContractEvents.Common
+  alias ContractEvents.Validators.ValidatorEpochPaymentDistributedEvent
 
   @celo_token_contract_symbols %{
     "stableToken" => "cUSD",
@@ -109,6 +118,67 @@ defmodule Explorer.Celo.Util do
     end)
     |> List.flatten()
     |> Enum.sort_by(& &1.epoch_number)
+    |> Enum.map_reduce(0, fn x, acc -> {x, acc + x.amount} end)
+  end
+
+  def fetch_and_structure_rewards(address_hash, from_date, to_date, fetch_for_validator_or_group) do
+    from_date =
+      case from_date do
+        nil -> ~U[2020-04-22 16:00:00.000000Z]
+        from_date -> from_date
+      end
+
+    to_date =
+      case to_date do
+        nil -> DateTime.utc_now()
+        to_date -> to_date
+      end
+
+    validator_epoch_payment_distributed = ValidatorEpochPaymentDistributedEvent.topic()
+    payment_param = fetch_for_validator_or_group <> "_payment"
+
+    query =
+      from(event in CeloContractEvent,
+        inner_join: block in Block,
+        on: event.block_number == block.number,
+        select: %{
+          amount: json_extract_path(event.params, [^payment_param]),
+          date: block.timestamp,
+          block_number: block.number,
+          block_hash: block.hash,
+          group: json_extract_path(event.params, ["group"]),
+          validator: json_extract_path(event.params, ["validator"])
+        },
+        order_by: [asc: block.number],
+        where: event.topic == ^validator_epoch_payment_distributed,
+        where: block.timestamp >= ^from_date,
+        where: block.timestamp < ^to_date
+      )
+
+    activated_votes_for_group =
+      cond do
+        fetch_for_validator_or_group == "validator" ->
+          query
+          |> CeloContractEvent.query_by_validator_param(address_hash)
+          |> Repo.all()
+
+        fetch_for_validator_or_group == "group" ->
+          query
+          |> CeloContractEvent.query_by_group_param(address_hash)
+          |> Repo.all()
+      end
+
+    activated_votes_for_group
+    |> Enum.map(fn x ->
+      Map.merge(
+        x,
+        %{
+          epoch_number: epoch_by_block_number(x.block_number),
+          group: Common.ca(x.group),
+          validator: Common.ca(x.validator)
+        }
+      )
+    end)
     |> Enum.map_reduce(0, fn x, acc -> {x, acc + x.amount} end)
   end
 end
